src/main/java/com/ranking/sample/SampleJsonController.java
package com.ranking.sample;

import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;


@RestController
@RequestMapping("/sample")
@Slf4j
public class SampleJsonController {

    @GetMapping(value = "/", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<String> sample() {

        log.info("Executing sample");
        return new ResponseEntity<>("{\"response\":\"sample\"}", HttpStatus.OK);
    }
}
